JVMCBC-315: Use double for ack threshold in Builder
Reviewed-on: http://review.couchbase.org/61924
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>

+++ src/main/java/com/couchbase/client/java/env/DefaultCouchbaseEnvironment.java
@@ -471,7 +471,7 @@ public class DefaultCouchbaseEnvironment extends DefaultCoreEnvironment implemen
         }
 
         @Override
-        public Builder dcpConnectionBufferAckThreshold(int dcpConnectionBufferAckThreshold) {
+        public Builder dcpConnectionBufferAckThreshold(double dcpConnectionBufferAckThreshold) {
             super.dcpConnectionBufferAckThreshold(dcpConnectionBufferAckThreshold);
             return this;
         }
